+++ test/src/main/java/com/trainee/services/ParentService.java
@@ -9,11 +9,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.trainee.models.FamilyMember;
 import com.trainee.models.Parent;
 import com.trainee.repositories.ParentRepository;
 
 /**
- * Clase donde se realizá la lógica de Parent para su respectivo CRUD
+ * Clase donde se realizá la lógica de la clase Parent para su respectivo CRUD
  * 
  * @author aescalan
  * @version 1.0
@@ -26,7 +27,7 @@ public class ParentService {
   private static final Logger log = LoggerFactory.getLogger(FamilyService.class);
 
   /**
-   * Método para obtener todos los Parents
+   * Método para obtener todos los objetos de la clase Parents
    * 
    * @return Todos los Parents
    */
@@ -37,9 +38,9 @@ public class ParentService {
   }
 
   /**
-   * Método para crear un Parent
+   * Método para crear un objeto de la clase Parent
    * 
-   * @param parent Recibir un Parent para luego guardar si se cumple la condición
+   * @param parent Recibir un objeto Parent para luego guardar si se cumple la condición
    * @return Manejo de HttpStatus según sea el caso
    */
   public ResponseEntity<Parent> postParent(Parent parent) {
@@ -56,9 +57,9 @@ public class ParentService {
   }
 
   /**
-   * Método para actualizar un Parent
+   * Método para actualizar un objeto de la clase Parent
    * 
-   * @param parent Recibir un Parent para luego actualizar si se cumple la
+   * @param parent Recibir un objeto Parent para luego actualizar si se cumple la
    *               condición
    * @return Manejo de HttpStatus según sea el caso
    */
@@ -75,14 +76,19 @@ public class ParentService {
   }
 
   /**
-   * Método para eliminar un Parent por id
+   * Método para eliminar un objeto de la clase Parent por id
    * 
    * @param id Id de Parent
    */
-  public void deleteParent(int id) {
-    log.info("Parent was deleted");
+  public ResponseEntity<Parent> deleteParent(int id) { 
     Parent parent = parentRepository.findById(id);
-    parentRepository.delete(parent);
+    if (parentRepository.findById(id) != null) {
+    	log.info("Parent was deleted");
+    	parentRepository.delete(parent);
+      return new ResponseEntity<Parent>(HttpStatus.ACCEPTED);
+    } else {
+      return new ResponseEntity<Parent>(HttpStatus.NOT_FOUND);
+    }
 
   }
 }
